:recycle: (pc-12879) rename PhoneError to PhoneFlip
